UnitData 클래스 작성
SO 분리
UnitData 클래스 작성
