+++ Assets/Script/ScriptableObject/UnitDataSO.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b8e23423deb8cb446bb1f44408564811
+guid: 85372292a323f1f4492fe2d270acc616
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
