+++ Assets/Script/Unit/BattleUnit.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class BattleUnit : Unit
+public class BattleUnit : DeckUnit
 {
     [SerializeField] private Team _team;
     public Team Team => _team;
@@ -106,10 +106,52 @@ public class BattleUnit : Unit
     public bool GetFlipX() => _renderer.flipX;
 
     public CutSceneType GetCutSceneType() => CutSceneType.center; // Skill 없어져서 바꿨어요
-    
-    public List<Vector2> GetAttackRange() => Data.GetAttackRange();
-    
-    public List<Vector2> GetMoveRange() => Data.GetMoveRange();
+
+    public List<Vector2> GetAttackRange()
+    {
+        List<Vector2> RangeList = new List<Vector2>();
+
+        int Acolumn = 11;
+        int Arow = 5;
+
+        for (int i = 0; i < Data.AttackRange.Length; i++)
+        {
+            if (Data.AttackRange[i])
+            {
+                int x = (i % Acolumn) - (Acolumn >> 1);
+                int y = (i / Acolumn) - (Arow >> 1);
+
+                Vector2 vec = new Vector2(x, y);
+
+                RangeList.Add(vec);
+            }
+        }
+
+        return RangeList;
+    }
+
+    public List<Vector2> GetMoveRange()
+    {
+        List<Vector2> RangeList = new List<Vector2>();
+
+        int Mrow = 5;
+        int Mcolumn = 5;
+
+        for (int i = 0; i < Data.MoveRange.Length; i++)
+        {
+            if (Data.MoveRange[i])
+            {
+                int x = (i % Mcolumn) - (Mcolumn >> 1);
+                int y = -((i / Mcolumn) - (Mrow >> 1));
+
+                Vector2 vec = new Vector2(x, y);
+
+                RangeList.Add(vec);
+            }
+        }
+
+        return RangeList;
+    }
 }
 
 // 22.02.16
