+++ Assets/Script/Unit/DeckUnit.cs
@@ -2,9 +2,9 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Unit : MonoBehaviour
+public class DeckUnit : MonoBehaviour
 {
-    public UnitData Data;
+    public UnitDataSO Data;
     
     [SerializeField] private Stat _stat;
     public Stat Stat => Data.RawStat; // Memo : 나중에 낙인, 버프 추가한 스탯으로 수정
